Добавлен модуль компонента навигации
